requirements.txt
APScheduler==3.6.3
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-InfluxDB==0.3.1
idna==2.9
influxdb==5.2.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paho-mqtt==1.5.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.14.0
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
